feat(nhctl): [install command] add install command : install helm applicaton
Former-commit-id: e947fdeafbe7284917d43a6d79b040cc5bb3c77f

+++ pkg/nhctl/cmd/port-forward.go
@@ -31,6 +31,10 @@ var portForwardCmd = &cobra.Command {
 			fmt.Println("error: please use -d to specify a kubernetes deployment")
 			return
 		}
+		if nameSpace == "" {
+			fmt.Println("error: please use -n to specify a kubernetes namespace")
+			return
+		}
 		// todo local port should be specificed ?
 		c := make(chan os.Signal)
 		signal.Notify(c, syscall.SIGHUP)  // kill -1
@@ -59,7 +63,7 @@ var portForwardCmd = &cobra.Command {
 			pid := os.Getpid()
 			ioutil.WriteFile(".pid", []byte(fmt.Sprintf("%d", pid)), 0644 )
 		}
-		err = kubectl.PortForward(ctx , kubeconfig, deployment, localPort, remotePort) // eg : ./utils/darwin/kubectl port-forward --address 0.0.0.0 deployment/coding  12345:22
+		err = kubectl.PortForward(ctx , kubeconfig, nameSpace, deployment, localPort, remotePort) // eg : ./utils/darwin/kubectl port-forward --address 0.0.0.0 deployment/coding  12345:22
 		if err != nil {
 			fmt.Printf("failed to forward port : %v\n", err)
 			err = os.Remove(".pid")
